// src/domain/enums/PostulacionEnums.enum.ts
export enum EstadoPostulacion {
  ENVIADA = 'enviada',
  EN_REVISION = 'en_revision',
  EN_PROCESO = 'en_proceso',
  ACEPTADA = 'aceptada',
  RECHAZADA = 'rechazada',
  RETIRADA = 'retirada',
}

export enum OrigenRecomendacion {
  SISTEMA = 'sistema',
  INTERMEDIADOR = 'intermediador',
}

export enum FrecuenciaAlerta {
  DIARIA = 'diaria',
  SEMANAL = 'semanal',
}